+++ .astro/settings.json
@@ -1,5 +1,5 @@
 {
 	"_variables": {
-		"lastUpdateCheck": 1714581921862
+		"lastUpdateCheck": 1715713277913
 	}
 }
