+++ README.md
@@ -23,12 +23,12 @@ File/Folder tree below:
 
   Example links:
 
-  - Bible = http://localhost:3000/WycliffeAssociates/en_ulb
-  - TN = http://localhost:3000/WycliffeAssociates/en_tn
-  - TQ = http://localhost:3000/WycliffeAssociates/en_tq
-  - BC = http://localhost:3000/WycliffeAssociates/en_bc
-  - TW = http://localhost:3000/WycliffeAssociates/en_tw
-  - TM = http://localhost:3000/WycliffeAssociates/en_tm
+  - Bible = http://localhost:4321/WycliffeAssociates/en_ulb
+  - TN = http://localhost:4321/WycliffeAssociates/en_tn
+  - TQ = http://localhost:4321/WycliffeAssociates/en_tq
+  - BC = http://localhost:4321/WycliffeAssociates/en_bc
+  - TW = http://localhost:4321/WycliffeAssociates/en_tw
+  - TM = http://localhost:4321/WycliffeAssociates/en_tm
 
 ## Folders
