+++ package.json
@@ -28,7 +28,7 @@
   },
   "dependencies": {
     "@astrojs/check": "^0.5.10",
-    "@astrojs/cloudflare": "9.1.0",
+    "@astrojs/cloudflare": "10.2.4",
     "@astrojs/solid-js": "4.1.0",
     "@astrojs/tailwind": "^5.1.0",
     "@kobalte/core": "^0.13.0",
@@ -36,10 +36,10 @@
     "astro": "^4.7.0",
     "fflate": "^0.8.2",
     "idb-keyval": "^6.2.1",
+    "logrocket": "^8.1.0",
     "p-limit": "^5.0.0",
     "solid-js": "^1.8.17",
-    "tailwindcss": "^3.4.3",
-    "vitest": "^1.5.3"
+    "tailwindcss": "^3.4.3"
   },
   "devDependencies": {
     "@cloudflare/workers-types": "^4.20240423.0",
@@ -49,10 +49,12 @@
     "@typescript-eslint/eslint-plugin": "^7.8.0",
     "@typescript-eslint/parser": "^7.8.0",
     "@vite-pwa/astro": "^0.4.0",
+    "vitest": "^1.5.3",
     "eslint": "^8.57.0",
     "eslint-plugin-astro": "^1.1.1",
     "eslint-plugin-jsx-a11y": "^6.8.0",
     "eslint-plugin-solid": "^0.14.0",
+    "htmlrewriter": "^0.0.7",
     "prettier": "3.2.5",
     "prettier-plugin-astro": "^0.13.0",
     "prettier-plugin-tailwindcss": "^0.5.14",
