+++ src/env.d.ts
@@ -3,3 +3,10 @@
 /// <reference types="astro/client" />
 /// <reference types="vite-plugin-pwa/info" />
 /// <reference types="vite-plugin-pwa/client" />
+/// <reference types="astro/client" />
+
+type Runtime = import("@astrojs/cloudflare").Runtime<Env>;
+
+declare namespace App {
+  interface Locals extends Runtime {}
+}
