+++ src/pages/api/repoIndex.ts
@@ -1,11 +1,12 @@
 import type { IcfEnv } from "@customTypes/types";
-import { getHeaders } from "functions/shared";
+import type { APIRoute } from "astro";
+import { getHeaders } from "@lib/api";
 
-export const onRequestGet: PagesFunction = async (context) => {
-  const request: Request = context.request;
-  const env = context.env as IcfEnv & typeof context.env;
+export const GET: APIRoute = async (context) => {
+  const runtime = context.locals.runtime;
+  const env = runtime.env as IcfEnv;
+  const { url } = context;
 
-  const url = new URL(request.url);
   const user = url.searchParams?.get("user");
   const repo = url.searchParams?.get("repo");
   if (!user || !repo || user == "undefined" || repo == "undefined") {
